+++ plugin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.0](https://github.com/paseo-cafe/paseo-cafe/compare/v0.1.1...v0.2.0) (2026-09-10)
+
+
+### Features
+
+* add agent-native catalog interfaces ([#70](https://github.com/paseo-cafe/paseo-cafe/issues/70)) ([f02edae](https://github.com/paseo-cafe/paseo-cafe/commit/f02edaef6c53e7abea3f21c49b74de24e21aff34))
+* improve Cafe discovery, trust, and plugin workflows ([#52](https://github.com/paseo-cafe/paseo-cafe/issues/52)) ([77337ab](https://github.com/paseo-cafe/paseo-cafe/commit/77337abf1bf7eb528581e4cb1d3c2ba4664260f5))
+* **plugin:** align native Cafe UI with website ([f376e5b](https://github.com/paseo-cafe/paseo-cafe/commit/f376e5bba6e24bb6d65c358b4c346f647e1b853c))
+
 ## [0.1.1](https://github.com/paseo-cafe/paseo-cafe/compare/v0.1.0...v0.1.1) (2026-09-10)